chore(plugin): split plugin::tables::vtable into submodules
We're going to add some alternate implementations, so let's not make
the lone .rs file too bloated
