+++ falco_plugin/src/lib.rs
@@ -64,7 +64,7 @@ pub mod base {
 ///     ExtractPlugin,
 ///     ExtractRequest,
 ///     field};
-/// use falco_plugin::tables::{TableReader, TablesInput};
+/// use falco_plugin::tables::TablesInput;
 ///
 /// struct MyExtractPlugin;
 /// impl Plugin for MyExtractPlugin {
@@ -639,8 +639,8 @@ pub mod listen {
 /// can use them from your plugin (e.g. in a separate thread) concurrently to other plugins
 /// (in the main thread).
 pub mod tables {
-    pub use crate::plugin::tables::vtable::TableReader;
-    pub use crate::plugin::tables::vtable::TableWriter;
+    pub use crate::plugin::tables::vtable::reader::TableReader;
+    pub use crate::plugin::tables::vtable::writer::TableWriter;
     pub use crate::plugin::tables::vtable::TablesInput;
 
     /// Exporting tables to other plugins
