+++ falco_plugin/src/plugin/tables/vtable/mod.rs
@@ -7,286 +7,26 @@ use crate::plugin::exported_tables::wrappers::{fields_vtable, reader_vtable, wri
 use crate::plugin::tables::data::Key;
 use crate::plugin::tables::table::raw::RawTable;
 use crate::plugin::tables::traits::{TableAccess, TableMetadata as ImportedTableMetadata};
-use crate::plugin::tables::vtable::TableError::BadVtable;
 use falco_plugin_api::{
-    ss_plugin_bool, ss_plugin_init_input, ss_plugin_owner_t, ss_plugin_rc, ss_plugin_state_data,
-    ss_plugin_state_type, ss_plugin_table_entry_t, ss_plugin_table_field_t,
-    ss_plugin_table_fieldinfo, ss_plugin_table_fields_vtable, ss_plugin_table_fields_vtable_ext,
-    ss_plugin_table_info, ss_plugin_table_input, ss_plugin_table_iterator_func_t,
-    ss_plugin_table_iterator_state_t, ss_plugin_table_reader_vtable,
-    ss_plugin_table_reader_vtable_ext, ss_plugin_table_t, ss_plugin_table_writer_vtable,
-    ss_plugin_table_writer_vtable_ext,
+    ss_plugin_init_input, ss_plugin_owner_t, ss_plugin_rc, ss_plugin_state_type,
+    ss_plugin_table_fields_vtable, ss_plugin_table_info, ss_plugin_table_input,
+    ss_plugin_table_reader_vtable, ss_plugin_table_t, ss_plugin_table_writer_vtable,
 };
 use std::ffi::CStr;
 use thiserror::Error;
+
+pub mod fields;
+pub mod reader;
+pub mod writer;
+
+use fields::TableFields;
+use reader::TableReader;
+use writer::TableWriter;
 
 #[derive(Error, Debug)]
 pub enum TableError {
     #[error("Missing entry {0} in table operations vtable")]
     BadVtable(&'static str),
-}
-
-/// A vtable containing table read access methods
-///
-/// It's used as a token to prove you're allowed to read tables in a particular context
-#[derive(Debug)]
-pub struct TableReader<'t> {
-    reader_ext: &'t ss_plugin_table_reader_vtable_ext,
-    pub(in crate::plugin::tables) last_error: LastError,
-}
-
-impl<'t> TableReader<'t> {
-    pub(crate) fn try_from(
-        reader_ext: &'t ss_plugin_table_reader_vtable_ext,
-        last_error: LastError,
-    ) -> Result<Self, TableError> {
-        Ok(TableReader {
-            reader_ext,
-            last_error,
-        })
-    }
-
-    pub(in crate::plugin::tables) fn get_table_name(
-        &self,
-        t: *mut ss_plugin_table_t,
-    ) -> Result<*const ::std::os::raw::c_char, TableError> {
-        Ok(unsafe {
-            self.reader_ext
-                .get_table_name
-                .ok_or(BadVtable("get_table_name"))?(t)
-        })
-    }
-
-    pub(in crate::plugin::tables) fn get_table_size(
-        &self,
-        t: *mut ss_plugin_table_t,
-    ) -> Result<u64, TableError> {
-        Ok(unsafe {
-            self.reader_ext
-                .get_table_size
-                .ok_or(BadVtable("get_table_size"))?(t)
-        })
-    }
-
-    pub(in crate::plugin::tables) fn get_table_entry(
-        &self,
-        t: *mut ss_plugin_table_t,
-        key: *const ss_plugin_state_data,
-    ) -> Result<*mut ss_plugin_table_entry_t, TableError> {
-        Ok(unsafe {
-            self.reader_ext
-                .get_table_entry
-                .ok_or(BadVtable("get_table_entry"))?(t, key)
-        })
-    }
-
-    pub(in crate::plugin::tables) fn read_entry_field(
-        &self,
-        t: *mut ss_plugin_table_t,
-        e: *mut ss_plugin_table_entry_t,
-        f: *const ss_plugin_table_field_t,
-        out: *mut ss_plugin_state_data,
-    ) -> Result<ss_plugin_rc, TableError> {
-        Ok(unsafe {
-            self.reader_ext
-                .read_entry_field
-                .ok_or(BadVtable("read_entry_field"))?(t, e, f, out)
-        })
-    }
-
-    pub(in crate::plugin::tables) fn release_table_entry_fn(
-        &self,
-    ) -> Option<
-        unsafe extern "C-unwind" fn(t: *mut ss_plugin_table_t, e: *mut ss_plugin_table_entry_t),
-    > {
-        self.reader_ext.release_table_entry
-    }
-
-    pub(in crate::plugin::tables) fn iterate_entries_fn(
-        &self,
-    ) -> Result<
-        unsafe extern "C-unwind" fn(
-            t: *mut ss_plugin_table_t,
-            it: ss_plugin_table_iterator_func_t,
-            s: *mut ss_plugin_table_iterator_state_t,
-        ) -> ss_plugin_bool,
-        TableError,
-    > {
-        self.reader_ext
-            .iterate_entries
-            .ok_or(BadVtable("iterate_entries"))
-    }
-}
-
-/// A vtable containing table write access methods
-///
-/// It's used as a token to prove you're allowed to write tables in a particular context
-#[derive(Debug)]
-pub struct TableWriter<'t> {
-    writer_ext: &'t ss_plugin_table_writer_vtable_ext,
-    pub(in crate::plugin::tables) last_error: LastError,
-}
-
-impl<'t> TableWriter<'t> {
-    pub(crate) fn try_from(
-        writer_ext: &'t ss_plugin_table_writer_vtable_ext,
-        last_error: LastError,
-    ) -> Result<Self, TableError> {
-        Ok(TableWriter {
-            writer_ext,
-            last_error,
-        })
-    }
-
-    pub(in crate::plugin::tables) fn clear_table(
-        &self,
-        t: *mut ss_plugin_table_t,
-    ) -> Result<ss_plugin_rc, TableError> {
-        unsafe {
-            Ok(self
-                .writer_ext
-                .clear_table
-                .ok_or(BadVtable("clear_table"))?(t))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn erase_table_entry(
-        &self,
-        t: *mut ss_plugin_table_t,
-        key: *const ss_plugin_state_data,
-    ) -> Result<ss_plugin_rc, TableError> {
-        unsafe {
-            Ok(self
-                .writer_ext
-                .erase_table_entry
-                .ok_or(BadVtable("erase_table_entry"))?(
-                t, key
-            ))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn create_table_entry(
-        &self,
-        t: *mut ss_plugin_table_t,
-    ) -> Result<*mut ss_plugin_table_entry_t, TableError> {
-        unsafe {
-            Ok(self
-                .writer_ext
-                .create_table_entry
-                .ok_or(BadVtable("create_table_entry"))?(t))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn destroy_table_entry(
-        &self,
-        t: *mut ss_plugin_table_t,
-        e: *mut ss_plugin_table_entry_t,
-    ) {
-        let Some(destroy_table_entry) = self.writer_ext.destroy_table_entry else {
-            return;
-        };
-
-        unsafe { destroy_table_entry(t, e) }
-    }
-
-    pub(in crate::plugin::tables) fn destroy_table_entry_fn(
-        &self,
-    ) -> Option<
-        unsafe extern "C-unwind" fn(t: *mut ss_plugin_table_t, e: *mut ss_plugin_table_entry_t),
-    > {
-        self.writer_ext.destroy_table_entry
-    }
-
-    pub(in crate::plugin::tables) fn add_table_entry(
-        &self,
-        t: *mut ss_plugin_table_t,
-        key: *const ss_plugin_state_data,
-        entry: *mut ss_plugin_table_entry_t,
-    ) -> Result<*mut ss_plugin_table_entry_t, TableError> {
-        unsafe {
-            Ok(self
-                .writer_ext
-                .add_table_entry
-                .ok_or(BadVtable("add_table_entry"))?(
-                t, key, entry
-            ))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn write_entry_field(
-        &self,
-        t: *mut ss_plugin_table_t,
-        e: *mut ss_plugin_table_entry_t,
-        f: *const ss_plugin_table_field_t,
-        in_: *const ss_plugin_state_data,
-    ) -> Result<ss_plugin_rc, TableError> {
-        unsafe {
-            Ok(self
-                .writer_ext
-                .write_entry_field
-                .ok_or(BadVtable("write_entry_field"))?(
-                t, e, f, in_
-            ))
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct TableFields<'t> {
-    fields_ext: &'t ss_plugin_table_fields_vtable_ext,
-}
-
-impl<'t> TableFields<'t> {
-    fn try_from(fields_ext: &'t ss_plugin_table_fields_vtable_ext) -> Result<Self, TableError> {
-        Ok(TableFields { fields_ext })
-    }
-
-    pub(in crate::plugin::tables) fn list_table_fields(
-        &self,
-        t: *mut ss_plugin_table_t,
-        nfields: *mut u32,
-    ) -> Result<*const ss_plugin_table_fieldinfo, TableError> {
-        unsafe {
-            Ok(self
-                .fields_ext
-                .list_table_fields
-                .ok_or(BadVtable("list_table_fields"))?(
-                t, nfields
-            ))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn get_table_field(
-        &self,
-        t: *mut ss_plugin_table_t,
-        name: *const ::std::os::raw::c_char,
-        data_type: ss_plugin_state_type,
-    ) -> Result<*mut ss_plugin_table_field_t, TableError> {
-        unsafe {
-            Ok(self
-                .fields_ext
-                .get_table_field
-                .ok_or(BadVtable("get_table_field"))?(
-                t, name, data_type
-            ))
-        }
-    }
-
-    pub(in crate::plugin::tables) fn add_table_field(
-        &self,
-        t: *mut ss_plugin_table_t,
-        name: *const ::std::os::raw::c_char,
-        data_type: ss_plugin_state_type,
-    ) -> Result<*mut ss_plugin_table_field_t, TableError> {
-        unsafe {
-            Ok(self
-                .fields_ext
-                .add_table_field
-                .ok_or(BadVtable("add_table_field"))?(
-                t, name, data_type
-            ))
-        }
-    }
 }
 
 #[derive(Debug)]
